pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "esuls"
version = "0.1.4"
description = "Utility library for async database operations, HTTP requests, and parallel execution"
readme = "README.md"
requires-python = ">=3.14"
authors = [
    {name = "IperGiove", email = "ipergiove@gmail.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

# All dependencies (always installed)
dependencies = [
    "aiosqlite>=0.21.0",
    "curl-cffi>=0.13.0",
    "fake-useragent>=2.2.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pillow>=12.0.0",
    "python-magic>=0.4.27",
]

[project.urls]
Homepage = "https://github.com/ipergiove/esuls"
Repository = "https://github.com/ipergiove/esuls"
